generate Fiat Shamir hints along fri inputs
